# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: FINN Licensing Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/licenses:
    get:
      operationId: listLicenses
      tags: [licenses]
      summary: List all licenses
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [licenses]
      summary: Create a license
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: License created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseCreated"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/licenses/{id}:
    get:
      operationId: getLicense
      tags: [licenses]
      summary: Get a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLicense
      tags: [licenses]
      summary: Update a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "200":
          description: License updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete a license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/licenses/{id}/toggle:
    post:
      operationId: toggleLicense
      tags: [licenses]
      summary: Toggle license status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        licenseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - licenseCount
        - createdAt
    ClientInput:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
...
    License:
      type: object
      properties:
        id:
          type: integer
        licenseKeyPreview:
          type: string
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        domain:
          type: string
        pluginAccess:
          type: string
          enum: [all, specific]
        productIds:
          type: array
          items:
            type: integer
          nullable: true
        status:
          type: string
          enum: [active, revoked]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - licenseKeyPreview
        - domain
        - pluginAccess
        - status
        - createdAt
    LicenseInput:
      type: object
      properties:
        clientId:
          type: integer
        domain:
          type: string
        pluginAccess:
          type: string
          enum: [all, specific]
        productIds:
          type: array
          items:
            type: integer
          nullable: true
        status:
          type: string
          enum: [active, revoked]
      required:
        - clientId
        - domain
        - pluginAccess
    LicenseCreated:
      type: object
      properties:
        license:
          $ref: "#/components/schemas/License"
        licenseKey:
          type: string
      required:
        - license
        - licenseKey
...