# lib/api-zod/src/generated/types/licenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FINN Licensing Server API
 * OpenAPI spec version: 1.0.0
 */
import type { LicenseInputPluginAccess } from "./licenseInputPluginAccess";
import type { LicenseInputStatus } from "./licenseInputStatus";

export interface LicenseInput {
  clientId: number;
  domain: string;
  pluginAccess: LicenseInputPluginAccess;
  productIds?: number[] | null;
  status?: LicenseInputStatus;
}